--- a/packages/common/src/formatters/__tests__/dateIsoFormatter.spec.ts
+++ b/packages/common/src/formatters/__tests__/dateIsoFormatter.spec.ts
@@ -16,7 +16,7 @@ describe('the Date ISO Formatter', () => {
 
   it('should provide a dateIso formatted input and return a formatted date value without time when valid date value is provided', () => {
     const value = '2019-05-03 00:00:01';
-    const result = Formatters.dateIso(0, 0, value, { input: 'dateIso' } as unknown as Column, {}, {} as any);
+    const result = Formatters.dateIso(0, 0, value, { type: 'dateIso' } as unknown as Column, {}, {} as any);
     expect(result).toBe('2019-05-03');
   });