--- a/packages/common/src/formatters/__tests__/dateTimeShortEuroFormatter.spec.ts
+++ b/packages/common/src/formatters/__tests__/dateTimeShortEuroFormatter.spec.ts
@@ -16,7 +16,7 @@ describe('the DateTimeShortEuro Formatter', () => {
 
   it('should provide a dateIso formatted input and return a formatted date value without time when valid date value is provided', () => {
     const value = '2019-05-01 02:36:07';
-    const result = Formatters.dateTimeShortEuro(0, 0, value, { input: 'dateIso' } as unknown as Column, {}, {} as any);
+    const result = Formatters.dateTimeShortEuro(0, 0, value, { type: 'dateIso' } as unknown as Column, {}, {} as any);
     expect(result).toBe('01/05/2019 02:36');
   });