--- a/packages/common/src/formatters/__tests__/dateTimeUsFormatter.spec.ts
+++ b/packages/common/src/formatters/__tests__/dateTimeUsFormatter.spec.ts
@@ -16,7 +16,7 @@ describe('the DateTimeUs Formatter', () => {
 
   it('should provide a dateIso formatted input and return a formatted date value without time when valid date value is provided', () => {
     const value = '2019-05-03 02:36:07';
-    const result = Formatters.dateTimeUs(0, 0, value, { input: 'dateIso' } as unknown as Column, {}, {} as any);
+    const result = Formatters.dateTimeUs(0, 0, value, { type: 'dateIso' } as unknown as Column, {}, {} as any);
     expect(result).toBe('05/03/2019 02:36:07');
   });